+++ .koppie/config/initializers/hyrax.rb
@@ -326,7 +326,7 @@ custom_queries = [Hyrax::CustomQueries::Navigators::CollectionMembers,
                   Hyrax::CustomQueries::FindFileMetadata,
                   Hyrax::CustomQueries::FindIdsByModel,
                   Hyrax::CustomQueries::FindManyByAlternateIds,
-                  Hyrax::CustomQueries::FindByReadGroup]
+                  Hyrax::CustomQueries::FindModelsByAccess]
 custom_queries.each do |handler|
   Hyrax.query_service.custom_queries.register_query_handler(handler)
 end
